% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstanmod1-package.R
\docType{package}
\name{rstanmod1-package}
\alias{rstanmod1-package}
\alias{rstanmod1}
\title{The 'rstanmod1' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. https://mc-stan.org
}
